Fix Orca to not generate unnecessary slices for plans with Subplans
Previously, Orca would generate unused extra slices when the plan
contained Subplans. This was due to a bug in the logic that set the
number of InitPlans (which do require an extra slice) to the number of
Subplans. This is incorrect--Orca doesn't generate InitPlans, so this
value can safely be set to 0. The dispatcher would create the slice
table and number of slices as "n = 1 + nMotions + nSubplans;" (where
nSubplans is actually the number of InitPlans).

For the query:

```

create table foo (a int, b int);

create table  bar (a int, b int);

test=# explain analyze select (select b from foo where b=bar.a) from bar;
                                                         QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice2; segments: 3)  (cost=0.00..1324039.51 rows=1 width=12) (actual time=2.762..2.762 rows=0 loops=1)
   ->  Result  (cost=0.00..1324039.51 rows=1 width=12) (never executed)
         ->  Seq Scan on bar  (cost=0.00..1324039.51 rows=334 width=4) (never executed)
         SubPlan 1  (slice2; segments: 3)
           ->  Result  (cost=0.00..431.00 rows=1 width=4) (never executed)
                 Filter: (foo.b = bar.a)
                 ->  Materialize  (cost=0.00..431.00 rows=1 width=4) (never executed)
                       ->  Broadcast Motion 3:3  (slice1; segments: 3)  (cost=0.00..431.00 rows=1 width=4) (never executed)
                             ->  Seq Scan on foo  (cost=0.00..431.00 rows=1 width=4) (never executed)
 Planning time: 92.378 ms
   (slice0)    Executor memory: 123K bytes.
   (slice1)    Executor memory: 58K bytes avg x 3 workers, 58K bytes max (seg0).
   (slice2)    Executor memory: 92K bytes avg x 3 workers, 92K bytes max (seg0).
   (slice3)
 Memory used:  128000kB
 Optimizer: Pivotal Optimizer (GPORCA)
 Execution time: 61.759 ms
(17 rows)

Time: 161.188 ms
test=#

```

We see that slice3 is initialized, but not used.

This doesn't apply to the Main branch, as this logic was removed in 93abe741c
and e7cc27fd05 .