Initialize attnumsWithEntries to prevent failures during CO rewrite
The rewrite process detects previous encodings of every column and stores them
in `attnumsWithEntries[]` to decide whether to add a new encoding or update the
existing one. That array was used uninitialized, which could result in "ERROR:
could not find tuple for attnum X for relid Y during scan on pg_attribute_encoding"
in some cases, especially with very wide tables.