Initialize attnumsWithEntries to prevent failures during CO rewrite (#6)
The rewrite process detects previous encodings of every column and stores them
in `attnumsWithEntries[]` to decide whether to add a new encoding or update the
existing one. Only the first 8 elements of that array were initialized, which could
result in "ERROR: could not find tuple for attnum X for relid Y during scan on
pg_attribute_encoding" in most cases, especially with wide tables.

## Here are some reminders before you submit the pull request
- [x] Add tests for the change
- [x] Pass `make installcheck`