Ignore "Extra Text" outputs in explain_format case
For instance in the Hash Join node, Extra Text is only printed when a
spill occurs, which heavily depends on the system state. Simply lowering
statement_mem does not reliably trigger a spill when the data set is
small.

So instead of checking Extra Text output in this case, we can skip
it, since we already have sufficient coverage in other tests.