Don't create motion on QE slice
The planner can execute some functions on segments. However, their contents
will also be planned on segments. Planning may create motions, which is
unacceptable on segments. Since the contents of functions may be unavailable
when planning the initial query (for example, a C-function with a call to SQL
in the SPI), it is sufficient to prevent motions from being created when
planning a function on a segment.

Ticket: ADBDEV-8689
(cherry picked from commit 50385e2ebc768a92cda0692a604df80981061512)